+++ examples/lib/stories/input/drag_callbacks_example.dart
@@ -3,14 +3,14 @@ import 'package:flame/events.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart' show Colors;
 
-class DraggablesExample extends FlameGame {
+class DragCallbacksExample extends FlameGame {
   static const String description = '''
     In this example we show you can use the `DragCallbacks` mixin on
     `PositionComponent`s. Drag around the Embers and see their position
     changing.
   ''';
 
-  DraggablesExample({required this.zoom});
+  DragCallbacksExample({required this.zoom});
 
   final double zoom;
   late final DraggableEmber square;
@@ -34,19 +34,12 @@ class DraggableEmber extends Ember with DragCallbacks {
   @override
   void update(double dt) {
     super.update(dt);
-    debugColor = isDragged && findGame() is DraggablesExample
-        ? Colors.greenAccent
-        : Colors.purple;
+    debugColor = isDragged ? Colors.greenAccent : Colors.purple;
   }
 
   @override
   void onDragUpdate(DragUpdateEvent event) {
-    if (findGame() is! DraggablesExample) {
-      event.continuePropagation = true;
-      return;
-    }
-
-    position.add(event.delta);
-    event.continuePropagation = false;
+    event.continuePropagation = true;
+    return;
   }
 }