+++ examples/lib/stories/input/overlapping_tap_callbacks_example.dart
@@ -4,7 +4,7 @@ import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 
-class OverlappingTappablesExample extends FlameGame {
+class OverlappingTapCallbacksExample extends FlameGame {
   static const String description = '''
     In this example we show you that events can choose to continue propagating
     to underlying components. The middle green square continue to propagate the
@@ -13,14 +13,19 @@ class OverlappingTappablesExample extends FlameGame {
 
   @override
   Future<void> onLoad() async {
-    add(TappableSquare(position: Vector2(100, 100)));
-    add(TappableSquare(position: Vector2(150, 150), continuePropagation: true));
-    add(TappableSquare(position: Vector2(100, 200)));
+    add(TapCallbacksSquare(position: Vector2(100, 100)));
+    add(
+      TapCallbacksSquare(
+        position: Vector2(150, 150),
+        continuePropagation: true,
+      ),
+    );
+    add(TapCallbacksSquare(position: Vector2(100, 200)));
   }
 }
 
-class TappableSquare extends RectangleComponent with TapCallbacks {
-  TappableSquare({Vector2? position, this.continuePropagation = false})
+class TapCallbacksSquare extends RectangleComponent with TapCallbacks {
+  TapCallbacksSquare({Vector2? position, this.continuePropagation = false})
       : super(
           position: position ?? Vector2.all(100),
           size: Vector2.all(100),