+++ packages/flame/lib/src/camera/viewport.dart
@@ -1,10 +1,7 @@
-import 'dart:ui';
-
+import 'package:flame/components.dart';
+import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
-import 'package:flame/src/anchor.dart';
-import 'package:flame/src/camera/camera_component.dart';
 import 'package:flame/src/camera/viewports/fixed_resolution_viewport.dart';
-import 'package:flame/src/components/core/component.dart';
 import 'package:flame/src/effects/provider_interfaces.dart';
 import 'package:meta/meta.dart';
 
@@ -22,11 +19,18 @@ import 'package:meta/meta.dart';
 /// A viewport establishes its own local coordinate system, with the origin at
 /// the top left corner of the viewport's bounding box.
 abstract class Viewport extends Component
-    implements AnchorProvider, PositionProvider, SizeProvider {
+    implements
+        AnchorProvider,
+        PositionProvider,
+        SizeProvider,
+        CoordinateTransform {
   Viewport({super.children});
 
   final Vector2 _size = Vector2.zero();
   bool _isInitialized = false;
+
+  @internal
+  final Transform2D transform = Transform2D();
 
   /// Position of the viewport's anchor in the parent's coordinate frame.
   ///
@@ -146,5 +150,17 @@ abstract class Viewport extends Component
     return (output?..setValues(x, y)) ?? Vector2(x, y);
   }
 
-  void transformCanvas(Canvas canvas) {}
+  @override
+  Vector2 parentToLocal(Vector2 point) {
+    return globalToLocal(point);
+  }
+
+  @override
+  Vector2 localToParent(Vector2 point) {
+    return localToGlobal(point);
+  }
+
+  void transformCanvas(Canvas canvas) {
+    canvas.transform2D(transform);
+  }
 }